requirements.txt
torch>=2.0
torchvision
wandb
numpy
scipy
einops
tqdm
pyyaml
